+++ binaries/pico_limbo/src/handlers/mod.rs
@@ -1,5 +1,4 @@
 pub mod configuration;
 pub mod handshake;
 pub mod login;
-pub mod play;
 pub mod status;