+++ binaries/pico_limbo/src/main.rs
@@ -10,7 +10,6 @@ use crate::configuration::config::{Config, ConfigError, load_or_create};
 use crate::handlers::configuration::on_acknowledge_finish_configuration;
 use crate::handlers::handshake::on_handshake;
 use crate::handlers::login::{on_custom_query_answer, on_login_acknowledged, on_login_start};
-use crate::handlers::play::on_player_position;
 use crate::handlers::status::{on_ping_request, on_status_request};
 use crate::server::network::Server;
 use crate::server_state::{ServerState, ServerStateBuildError};
@@ -48,7 +47,6 @@ async fn main() -> ExitCode {
         .on(on_login_start)
         .on(on_login_acknowledged)
         .on(on_custom_query_answer)
-        .on(on_player_position)
         .on(on_acknowledge_finish_configuration)
         .run()
         .await;