+++ binaries/pico_limbo/src/server/mod.rs
@@ -1,5 +1,6 @@
 pub mod client;
 mod client_inner;
+mod controllable_interval;
 pub mod event_handler;
 pub mod game_profile;
 mod named_packet;