+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Keep alive packet not properly sending for 1.7.x
+
+### Removed
+
+- Unused `PlayerPositionPacket`
+
 ## [1.3.1+mc1.21.8] - 2025-07-19
 
 ### Changed